reuse initDemoPanCamera
